requirements.txt
customtkinter==5.2.1
mysql-connector-python==8.2.0
